Inquis Shuffling
A number of changes to the Inquisition, because I hated how inflated some of it was, and nonsensical the rest was. At least to my smooth brain.

So, getting into it, there's a few simple changes, and some that I'll need to justify(I won't.).

All Ortho statblocks are now a flat +8.
Inquis is a flat +10.
Those statblock changes are NERFS, except in the case of Adjudicator, who gets some much needed love.

Silver blessed was made standard, given they really should have it.

Adjudicator:
 - Slip arrival marques increased from 4 to 5.
 - Given Silver Blessed.
 - +1CON/-1SPD
 - Wrestling from 2 to 4.
 - Climbing from 1 to 3.(Who thought this was a good idea?)
 - Athletics from 3 to 4.
 - Devotion cap from T1 to T2, to be in line with the tier. This puts them above Disciples as miracle casters, given Disciples retain the T1 cap, though still have their T2 miracles.
 - Spear loadout given a GWS.

Disciple:
 - Given Silver Blessed.
 - -1STR.

Inquisitor:
 - Puritan lost Medium Armour training. They already had DE.
 - Puritan and Ordinator statspread adjusted. 10 max, in line with the Knight Captain.
    - Puritan: -2PER, -2INT, -2STR, +1SPD, +1END. For a total of 10 points, with 3CON/END, 2SPD, 1PER/INT.
    - Ordinator: -1INT/PER. For a total of 10 points, with 3CON/END, 2STR, 1PER/INT.
 - Fixed a silly little 'mistake' that someone left with an Inq loadout getting an extra slip. Wow. Wonder how that happened.

Otherwise, removed some vestigial stuff.